+++ libs/ymir-core/src/ymir/media/loader/loader_chd.cpp
@@ -11,6 +11,7 @@
 
 #include <libchdr/chd.h>
 
+#include <charconv>
 #include <span>
 #include <string>
 #include <utility>
@@ -305,6 +306,9 @@ bool Load(std::filesystem::path chdPath, Disc &disc, bool preloadToRAM) {
             track.startFrameAddress = frameAddress;
             track.endFrameAddress = frameAddress + frames - 1;
             track.interleavedSubchannel = false;
+            auto &index = track.indices.emplace_back();
+            index.startFrameAddress = track.startFrameAddress;
+            index.endFrameAddress = track.endFrameAddress;
             frameAddress += frames;
             byteOffset += frames * track.sectorSize;